chore: Move types back to its components
