--- src/types/types.tsx
@@ -1,12 +1,0 @@
-import { ReactNode } from 'react'
-export interface AppBarWebProps {
-  title: string
-  version: string
-}
-export interface AppBarMobileProps {
-  version: string
-}
-export interface MainProps {
-  version: string
-  children: ReactNode
-}
